+++ apps/gui/src/controllers/MainController.h
@@ -288,9 +288,9 @@ public slots:
 	// forward calls to ScaledSpectrum of FFTAnalyzer
 	// see ScaledSpectrum.h for documentation
 	qreal getFftGain() const { return m_fft.getScaledSpectrum().getGain(); }
-	void setFftGain(const qreal& value) { m_fft.getScaledSpectrum().setGain(value); emit gainChanged(); emit presetChanged(); }
+	void setFftGain(const qreal& value) { m_fft.getScaledSpectrum().setGain(static_cast<float>(value)); emit gainChanged(); emit presetChanged(); }
 	qreal getFftCompression() const { return m_fft.getScaledSpectrum().getCompression(); }
-	void setFftCompression(const qreal& value) { m_fft.getScaledSpectrum().setCompression(value); emit compressionChanged(); emit presetChanged(); }
+	void setFftCompression(const qreal& value) { m_fft.getScaledSpectrum().setCompression(static_cast<float>(value)); emit compressionChanged(); emit presetChanged(); }
 	bool getDecibelConversion() const { return m_fft.getScaledSpectrum().getDecibelConversion(); }
 	void setDecibelConversion(bool value) { m_fft.getScaledSpectrum().setDecibelConversion(value); emit decibelConversionChanged(); emit presetChanged(); }
 	bool getAgcEnabled() const { return m_fft.getScaledSpectrum().getAgcEnabled(); }
