+++ libs/sound2osc-core/src/dsp/FFTAnalyzer.cpp
@@ -46,7 +46,7 @@ void FFTAnalyzer::calculateWindow()
 	// Hann Window function
 	// used to prepare the PCM data for FFT
 	for (int i=0; i<NUM_SAMPLES; ++i) {
-		m_window[i] = 0.5f * (1 - qCos((2 * M_PI * i) / (NUM_SAMPLES - 1)));
+		m_window[i] = 0.5f * (1 - static_cast<float>(qCos((2 * M_PI * i) / (NUM_SAMPLES - 1))));
 	}
 }
 
@@ -55,7 +55,7 @@ void FFTAnalyzer::calculateFFT(bool lowSoloMode)
 	// apply window:
 	int bufferOffset = m_inputBuffer.getCapacity() - NUM_SAMPLES;
 	for (int i=0; i < NUM_SAMPLES; ++i) {
-		m_buffer[i] = m_inputBuffer.at(i + bufferOffset) * m_window[i];
+		m_buffer[i] = static_cast<float>(m_inputBuffer.at(i + bufferOffset)) * m_window[i];
 	}
 
 	// apply FFT:
